tests/wpt/tests/html/dom/render-blocking/element-render-blocking-006.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Adding blocking attr in the head has an effect</title>

<link id=link rel=expect href="#last">
<script>
link.blocking = "render"

async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_true(!!document.getElementById("last")));
    t.done();
  });
}, "adding 'blocking=render' in the head makes it blocking");
</script>
</head>
<body>
  <div id="first"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="second"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="last"></div>
</body>
